Fix build compatibility issues for Next 15
