+++ next.config.ts
@@ -1,13 +1,14 @@
 import type { NextConfig } from "next";
+import type { RemotePattern } from 'next/dist/shared/lib/image-config';
 
 const isStaticExport = process.env.STATIC_EXPORT === 'true'
 
-const remotePatterns = [
-  { protocol: 'https', hostname: 'images.unsplash.com' },
-  { protocol: 'https', hostname: 'i.ytimg.com' },
-  { protocol: 'https', hostname: 'img.youtube.com' },
-  { protocol: 'https', hostname: 'avatars.githubusercontent.com' },
-  { protocol: 'https', hostname: 'lh3.googleusercontent.com' },
+const remotePatterns: RemotePattern[] = [
+  { protocol: 'https', hostname: 'images.unsplash.com', pathname: '/**' },
+  { protocol: 'https', hostname: 'i.ytimg.com', pathname: '/**' },
+  { protocol: 'https', hostname: 'img.youtube.com', pathname: '/**' },
+  { protocol: 'https', hostname: 'avatars.githubusercontent.com', pathname: '/**' },
+  { protocol: 'https', hostname: 'lh3.googleusercontent.com', pathname: '/**' },
 ]
 
 const nextConfig: NextConfig = {
