+++ src/components/theme-provider.tsx
@@ -2,8 +2,10 @@
 "use client"
 
 import * as React from "react"
-import { ThemeProvider as NextThemesProvider } from "next-themes"
-import { type ThemeProviderProps } from "next-themes-dist-types"
+import {
+  ThemeProvider as NextThemesProvider,
+  type ThemeProviderProps,
+} from "next-themes"
 
 export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
   // Use suppressHydrationWarning to prevent hydration mismatch errors
