+++ src/components/ui/NewMarkdownRenderer.tsx
@@ -2,6 +2,7 @@
 
 import React from 'react';
 import ReactMarkdown from 'react-markdown';
+import type { ComponentPropsWithoutRef } from 'react';
 import remarkGfm from 'remark-gfm';
 import rehypeRaw from 'rehype-raw';
 import { NewCodeBlock } from './NewCodeBlock';
@@ -32,7 +33,9 @@ export function NewMarkdownRenderer({ content }: MarkdownRendererProps) {
         remarkPlugins={[remarkGfm]}
         rehypePlugins={[rehypeRaw]}
         components={{
-          code({ inline, className, children, ...props }) {
+          code({ inline, className, children, ...props }: ComponentPropsWithoutRef<'code'> & {
+            inline?: boolean;
+          }) {
             const match = /language-(\w+)/.exec(className || '');
             const language = match ? match[1] : '';
 
@@ -52,7 +55,9 @@ export function NewMarkdownRenderer({ content }: MarkdownRendererProps) {
             );
           },
           // Custom component for YouTube embeds
-          div({ className, ...props }) {
+          div({ className, ...props }: ComponentPropsWithoutRef<'div'> & {
+            'data-video-id'?: string;
+          }) {
             if (className === 'youtube-embed' && props['data-video-id']) {
               return <NewVideoEmbed videoId={props['data-video-id']} />;
             }
