+++ src/lib/supabase/client.ts
@@ -1,7 +1,5 @@
 import { createBrowserClient as createSupabaseBrowserClient } from '@supabase/ssr'
 
-type ClientOptions = Parameters<typeof createSupabaseBrowserClient>[2]
-
 const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL
 const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
 
@@ -13,12 +11,5 @@ if (!supabaseAnonKey) {
   throw new Error('Missing NEXT_PUBLIC_SUPABASE_ANON_KEY environment variable.')
 }
 
-const defaultOptions: ClientOptions = {
-  auth: {
-    persistSession: true,
-    autoRefreshToken: true,
-  },
-}
-
 export const createBrowserClient = () =>
-  createSupabaseBrowserClient(supabaseUrl, supabaseAnonKey, defaultOptions)
+  createSupabaseBrowserClient(supabaseUrl, supabaseAnonKey)
